refactor: fsd 구조에 맞게 도메인 마다 분리 진행
