+++ src/features/react-query-generate/module/queryGenerator.ts
@@ -1,6 +1,6 @@
 import { Method } from "axios";
-import { Parameters } from "../api/docs";
-import { typeConverter } from "./typeConverter";
+import { Parameters } from "../../../pages/popup/shared/api/docs";
+import { typeConverter } from "../../../pages/popup/shared/util/typeConverter";
 
 const generateQueryKey = (params: Parameters[]) => {
   return params
