+++ src/pages/popup/pages/ApiList/ui/ApiList.tsx
@@ -5,19 +5,19 @@ import { FiMenu as MenuIcon } from "react-icons/fi";
 import useHandleAuth from "@/features/auth/hooks/useHandleAuth";
 import AuthModal from "@/features/auth/ui/auth-modal/AuthModal";
 import useSearch from "@/features/search/module/hooks/useSearch";
-import useHandleSetting from "@/pages/popup/hooks/Popup/useHandleSetting";
+import useHandleSetting from "@/features/setting/module/hooks/useHandleSetting";
 import Search from "@/pages/popup/ui/Search";
 import SettingDrawer from "@/pages/popup/ui/SettingDrawer";
 import SettingModal from "@/pages/popup/ui/SettingModal";
-import BlankApi from "@/pages/popup/ui/error/BlankApi";
-import Loading from "@/pages/popup/ui/loading/Loading";
-import useHandlePopup from "../hooks/useHandlePopup";
-import { popupStyle } from "./popup.css";
+import Loading from "@/shared/ui/Loading/Loading";
+import BlankApi from "@/widgets/api-list/ui/blank/BlankApi";
 import ApiList from "@/widgets/api-list/ui/normal-list/ApiList";
+import { useHandleApiList } from "../../../../../widgets/api-list/module/hooks/useHandleApiList";
+import { apiListStyle } from "./apiList.css";
 
 const Popup = () => {
   const { loading, apiList, filteredApiList, onClickAPI, setFilteredApiList } =
-    useHandlePopup();
+    useHandleApiList();
 
   const { search, onChange } = useSearch({ apiList, setFilteredApiList });
   const { authorized, onChangeAuth, onSaveAuth } = useHandleAuth();
@@ -27,13 +27,13 @@ const Popup = () => {
   const { open, openDrawer, closeDrawer } = useDrawer();
 
   return (
-    <div id="main" className={popupStyle.app}>
-      <header className={popupStyle.header}>
-        <div className={popupStyle.settingWrapper}>
-          <button className={popupStyle.menuButton} onClick={openDrawer}>
+    <div id="main" className={apiListStyle.app}>
+      <header className={apiListStyle.header}>
+        <div className={apiListStyle.settingWrapper}>
+          <button className={apiListStyle.menuButton} onClick={openDrawer}>
             <MenuIcon size={25} color={vars.color.white} />
           </button>
-          <div className={popupStyle.settingButtonWrapper}>
+          <div className={apiListStyle.settingButtonWrapper}>
             <SettingModal
               withReactQuery={withReactQuery}
               toggleReactQuery={toggleReactQuery}
@@ -47,7 +47,7 @@ const Popup = () => {
           </div>
         </div>
       </header>
-      <div className={popupStyle.searchWrapper}>
+      <div className={apiListStyle.searchWrapper}>
         <Search value={search} onChange={onChange} />
       </div>
       {!!loading && <Loading />}
