+++ src/pages/popup/pages/ApiList/ui/apiList.css.ts
@@ -1,7 +1,7 @@
 import { vars } from "@/shared/ui/styles/theme.css";
 import { style } from "@vanilla-extract/css";
 
-export const popupStyle = {
+export const apiListStyle = {
   app: style({
     position: "absolute",
     backgroundColor: vars.color.background,
