+++ src/pages/popup/ui/CodeBlockView.tsx
@@ -5,7 +5,7 @@ import { ForwardRefRenderFunction, forwardRef } from "react";
 import { IoMdClose as CloseIcon } from "react-icons/io";
 import { IoChevronBackOutline as BackIcon } from "react-icons/io5";
 import { Mode } from "../pages/Request/Request";
-import { requestStyle } from "../pages/Request/request.css";
+import { requestStyles } from "../pages/Request/request.css";
 import CodeBlock from "./CodeBlock";
 
 interface CodeBlockProps {
@@ -44,11 +44,11 @@ const CodeBlockView: ForwardRefRenderFunction<
   return (
     <div
       style={{ width: isModal ? "80vw" : "auto" }}
-      className={requestStyle.modal}
+      className={requestStyles.modal}
     >
-      <div className={requestStyle.response}>
-        <div className={requestStyle.descriptionWrapper}>
-          <div className={requestStyle.leftWrapper}>
+      <div className={requestStyles.response}>
+        <div className={requestStyles.descriptionWrapper}>
+          <div className={requestStyles.leftWrapper}>
             {!!isModal && (
               <>
                 <button
@@ -56,7 +56,7 @@ const CodeBlockView: ForwardRefRenderFunction<
                     mode === "RESPONSE" && onClose && onClose();
                     mode !== "RESPONSE" && onClickBack && onClickBack();
                   }}
-                  className={requestStyle.iconButton}
+                  className={requestStyles.iconButton}
                   type="button"
                 >
                   {mode === "RESPONSE" && (
@@ -68,7 +68,7 @@ const CodeBlockView: ForwardRefRenderFunction<
                 </button>
 
                 <h3
-                  className={requestStyle.requestDescription}
+                  className={requestStyles.requestDescription}
                   style={{
                     color: vars.color[descriptionColor],
                   }}
@@ -83,7 +83,7 @@ const CodeBlockView: ForwardRefRenderFunction<
                   onClick={() => {
                     onClickBack && onClickBack();
                   }}
-                  className={requestStyle.iconButton}
+                  className={requestStyles.iconButton}
                   type="button"
                 >
                   <BackIcon size={24} color={vars.color.white} />
@@ -92,7 +92,7 @@ const CodeBlockView: ForwardRefRenderFunction<
             )}
             {!isModal && (
               <h3
-                className={requestStyle.requestDescription}
+                className={requestStyles.requestDescription}
                 style={{
                   color: vars.color[descriptionColor],
                 }}
@@ -101,7 +101,7 @@ const CodeBlockView: ForwardRefRenderFunction<
               </h3>
             )}
           </div>
-          <div className={requestStyle.buttonWrapper}>
+          <div className={requestStyles.buttonWrapper}>
             {onClickCopy && (
               <Button color="green" onClick={onClickCopy}>
                 COPY
