+++ src/widgets/api-list/ui/blank/BlankApi.tsx
