+++ src/widgets/api-list/ui/darg-list/DragApiList.tsx
@@ -2,8 +2,8 @@ import DragAndDrop from "@/shared/ui/DragAndDrop";
 import { Dispatch } from "react";
 import { IoMdMenu as MenuIcon } from "react-icons/io";
 import { APIWithOrder } from "../../../../entities/sequence/model/sequence-store";
-import APIItem from "../../../../pages/popup/ui/APIItem/APIItem";
-import { dragAPIStyles } from "./dragApi.css";
+import APIItem from "../api-item/ApiItem";
+import { dragApiStyles } from "./dragApi.css";
 
 interface ApiListProps {
   ApiList: APIWithOrder[];
@@ -15,7 +15,7 @@ const DragApiList = ({ ApiList, setApiList }: ApiListProps) => {
     <DragAndDrop
       itemList={ApiList}
       renderDragItem={(item) => (
-        <div className={dragAPIStyles.wrapper}>
+        <div className={dragApiStyles.wrapper}>
           <APIItem key={item.key} api={item.api} />
           <MenuIcon size={24} color="white" />
         </div>
