+++ src/widgets/api-list/ui/normal-list/ApiList.tsx
@@ -1,6 +1,6 @@
 import { API, ApiList } from "@/pages/content/modules/getApiList";
 import { vars } from "@/shared/ui/styles/theme.css";
-import APIItem from "../../../../pages/popup/ui/APIItem/APIItem";
+import APIItem from "../api-item/ApiItem";
 import { ApiListStyle } from "../apiList.css";
 
 interface ApiListProps {
