+++ src/widgets/api-list/ui/toggle-list/ToggleSwaggerApiList.tsx
@@ -1,7 +1,7 @@
 import { API, ApiList } from "@/pages/content/modules/getApiList";
-import APIItem from "@/pages/popup/ui/APIItem/APIItem";
 import CheckBox from "@/shared/ui/CheckBox";
 import { vars } from "@/shared/ui/styles/theme.css";
+import APIItem from "@/widgets/api-list/ui/api-item/ApiItem";
 import { APIWithOrder } from "../../../../entities/sequence/model/sequence-store";
 import { ApiListStyle } from "../apiList.css";
